chore: add Input/Output types to imports
